DistriMule: Allow not assigning station command to trader
When a trader is assigned to a station commander it becomes impassible
to modify the order without removing the assignment and issuing a new
order.

Not setting the station as commander makes modifying the order much
simpler. This is already possible for the StationMule, so why not also
make it possible for DistriMule?

It defaults to "Assign Ship" to true, to keep the current behaviour.